fix STT hook from daemon
